Move real time logs into their own file
